멀쩡한 사각형 재풀이(long * long을 해야 답이나왔음, 너무 큰 수는 오버플로우 나왔기때문에)

+++ Algorithm/Programmers/Level2/Java/Plain_Square/Plain_Square.java
@@ -11,7 +11,7 @@ public class Plain_Square {
 	}
 
 	public static long solution(int w, int h) {
-		long answer = w * h;
+		long answer = (long)w * (long)h;
         long gcd_v = gcd(w,h);
         answer = answer - (gcd_v * ((w/gcd_v) + (h/gcd_v) -1));
